drop custom setting semanticRoute for PDP, PLP and brand page
